Refactor: Extract other commands
