+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+export { start } from "./start.command";
+export { shutdown } from "./shutdown.command";
+export { status } from "./status.command";
+export { restart } from "./restart.command";
+export { logs } from "./logs.command";
