+++ src/commands/restart.command.ts
@@ -1,0 +1,13 @@
+import { createConnection } from "net";
+
+export const restart = async (name: string) => {
+  const client = createConnection(".pm.sock")
+  client.on('connect', () => {
+    client.write("restart " + name);
+  });
+  client.on("data", (data) => {
+    //TODO handle stderr/stdin and also write on stderr
+    process.stdout.write(data);
+  })
+}
+
