+++ src/index.ts
@@ -1,7 +1,5 @@
-import { spawn, ChildProcess } from "child_process";
-import { Socket, createServer, createConnection } from "net";
 import { Command } from 'commander';
-import { start } from "./commands/start.command";
+import { start, shutdown, status, logs, restart } from "./commands";
 import eco from "./ecosystem.config";
 
 const program = new Command();
@@ -31,51 +29,3 @@ program.command("shutdown")
   })
 
 program.parse(process.argv);
-
-async function shutdown() {
-  const client = createConnection(".pm.sock")
-  client.on('connect', () => {
-    client.write("shutdown");
-  });
-}
-
-async function logs(name?: string) {
-  const client = createConnection(".pm.sock")
-  client.on('connect', () => {
-    if (name) {
-      client.write("logs " + name);
-    } else {
-      //all logs
-      client.write("logs");
-    }
-
-  });
-  client.on("data", (data) => {
-    //TODO handle stderr/stdin and also write on stderr
-    process.stdout.write(data);
-  })
-}
-
-
-
-async function restart(name: string) {
-  const client = createConnection(".pm.sock")
-  client.on('connect', () => {
-    client.write("restart " + name);
-  });
-  client.on("data", (data) => {
-    //TODO handle stderr/stdin and also write on stderr
-    process.stdout.write(data);
-  })
-}
-
-
-async function status() {
-  const client = createConnection(".pm.sock")
-  client.on('connect', () => {
-    client.write("status");
-  });
-  client.on("data", (data) => {
-    console.log(data.toString());
-  })
-}
